requirements.txt
fastapi==0.115.6
python-dotenv==1.0.1
python-telegram-bot==21.9
hypercorn==0.17.3
requests==2.31.0
sqlite-utils==3.36
